Add files via upload
Contains the allocated csv file for the data integration part along with the project files. 

Create a sysintg folder in your htdocs folder of xampp and then extract each of those rar files into the folder. (Did this instead of one project archive because GitHub has limitations or restrictions in file size in uploading)